+++ src/main/java/com/spd/baraholka/image/service/ImageService.java
@@ -8,9 +8,9 @@ import java.util.*;
 public interface ImageService {
     ImageResource save(ImageResource imageResource);
 
-    void saveImageResources(long adId, List<ImageResource> imageResources);
+    void saveImageResources(int adId, List<ImageResource> imageResources);
 
-    Image uploadImage(long adId, MultipartFile image);
+    Image uploadImage(int adId, MultipartFile image);
 
     List<Image> getAllUnattached();
 
@@ -20,11 +20,11 @@ public interface ImageService {
 
     List<ImageResource> getPrimary(List<Integer> adIds);
 
-    List<ImageResource> getAllByAdId(long adId);
+    List<ImageResource> getAllByAdId(int adId);
 
-    void deleteAllByAdId(long adId);
+    void deleteAllByAdId(int adId);
 
-    void deleteImage(long imageId);
+    void deleteImage(int imageId);
 
     void deleteImage(String imageUrl);
